+++ com.chartboost.mediation.demo/Editor/DemoSetupOnEditorLaunch.cs
@@ -21,7 +21,7 @@ namespace Chartboost.Editor
       [MenuItem("Chartboost Mediation/Setup Demo")]
       private static void SetupDemoApp()
       {
-         Debug.Log("Configuring Chartboost Mediation SDK Unity Demo App.");
+         Debug.Log($"Configuring {ProductName}.");
          
          PlayerSettings.companyName = CompanyName;
          PlayerSettings.productName = ProductName;
@@ -45,7 +45,6 @@ namespace Chartboost.Editor
          PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Android, ApplicationBundleIdentifier);
          PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.iOS, ApplicationBundleIdentifier);
          PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Standalone, ApplicationBundleIdentifier);
-         PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Unknown, ApplicationBundleIdentifier);
 
          PlayerSettings.SetAdditionalCompilerArgumentsForGroup(BuildTargetGroup.Android, CompilerFlags);
          PlayerSettings.SetAdditionalCompilerArgumentsForGroup(BuildTargetGroup.iOS, CompilerFlags);